Uploading homework 02
Completed the homework python file
Added the notebook file, a bit uncomplete because couldn't run the whole python file again, due to changes on the containers.
Added the modified docker-compose.yaml file to save the downloaded weights on the local folder
Added the Dockerfile to add the downloaded weights to the new image "ollama-gemma2b"
Did not pushed the weights because the size is not allowed.
probably didn't need to add the full ollama_files folder, but just in case.
Added some screenshots of my work. I worked only with Podman